backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "fastapi-mail>=1.5.0",
    "fastapi-pagination>=0.14.1",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "starlette-babel>=1.0.3",
    "sqladmin>=0.21.0",
    "itsdangerous>=2.2.0",
    "boto3>=1.35.0",
    "python-multipart>=0.0.12",
    "pillow>=10.4.0",
    "magic-filter>=0.0.8",
    "aioboto3>=15.1.0",
    "markitdown[docx,pdf,pptx,xls,xlsx]>=0.1.3",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "langchain-anthropic>=1.0.1",
    "langchain-chroma>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "tiktoken>=0.11.0",
    "chromadb>=0.5.4",
    "dashscope>=1.24.5",
    "langchain-community>=0.4.1",
    "google-genai>=1.52.0",
    "cryptography>=45.0.7",
    "redis>=6.4.0",
    "authlib>=1.6.4",
    "asyncpg>=0.30.0",
    "sqlalchemy-utils>=0.42.0",
    "mcp>=1.18.0",
    "stripe>=11.3.0",
    "langchain-xai>=1.0.0",
    "langchain-google-genai>=3.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
]
